+++ std/crypto/sha1_test.ts
